Rename app.js to package.json, remove glob imports
